shefeels-backend/requirements.txt
fastapi[all]
sqlalchemy
sqlmodel
alembic
asyncpg
# stripe (removed) - replaced with PayGate integration (no python package required)
openai
python-multipart
passlib
pyjwt
uvicorn[standard]
httpx
pytest
pytest-asyncio
pytest-mock
websockets
python-dotenv
posthog
psycopg2
python-jose
pillow
aiosmtplib
boto3
aiohttp
tiktoken
redis
bcrypt>=4.1.0,<5.0.0
passlib[bcrypt]>=1.7.4
google-auth
requests
SpeechRecognition
pydub
together
elevenlabs
